Added some sounds, unity scene mods
Added a spear throwing and spear hitting sound. Spear hit should play
when it collides with anything. Spear throw should play when it's
thrown. Added a speech to each of the end screens. Added an animation to
the riders of the other chariots so it looks realistic. If the AI is
jabbing, the way to get the animation is
getComponent<Animation>().Play("jab");, if the rig on the model is
legacy. It should be in the 3D models folder, as jabbing-->rigging.